Fix type conversion warnings in OpenRecentController and PBGitGrapher
- Fix implicit conversion from NSUInteger to int in OpenRecentController.m:114
- Fix implicit conversion from NSInteger to int in OpenRecentController.m:149
- Fix implicit conversion from size_type to int in PBGitGrapher.mm:96
- Add explicit type casting to resolve compiler warnings safely

+++ Classes/git/PBGitGrapher.mm
@@ -93,7 +93,7 @@ void add_line(struct PBGitGraphLine *lines, int *nLines, int upper, int from, in
 					didFirst = YES;
 					currentLanes->push_back(*it);
 					currentLane = currentLanes->back();
-					newPos = currentLanes->size();
+					newPos = (int)currentLanes->size();
 					add_line(lines, &currentLine, 1, i, newPos,(*it)->index());
 					if (nParents)
 						add_line(lines, &currentLine, 0, newPos, newPos,(*it)->index());
